requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.0  # Updated for better compatibility
scikit-learn==1.6.1  # Updated to match your pickle file
joblib==1.3.2
opencv-python==4.8.1.78
pillow==10.1.0
easyocr==1.7.0
matplotlib==3.8.2
requests==2.31.0